Extract settings views
